+++ framework/model/src/test/java/is/codion/framework/model/DefaultEntityConditionModelTest.java
@@ -48,7 +48,7 @@ public class DefaultEntityConditionModelTest {
 					.build();
 
 	private final EntityConditionModel conditionModel = new DefaultEntityConditionModel(Employee.TYPE,
-					CONNECTION_PROVIDER, new EntityColumnConditionModelFactory(CONNECTION_PROVIDER));
+					CONNECTION_PROVIDER, new EntityConditionModelFactory(CONNECTION_PROVIDER));
 
 	@Test
 	void test() {
@@ -65,7 +65,7 @@ public class DefaultEntityConditionModelTest {
 	@Test
 	void noSearchColumnsDefined() {
 		EntityConditionModel model = new DefaultEntityConditionModel(Detail.TYPE,
-						CONNECTION_PROVIDER, new EntityColumnConditionModelFactory(CONNECTION_PROVIDER));
+						CONNECTION_PROVIDER, new EntityConditionModelFactory(CONNECTION_PROVIDER));
 		//no search columns defined for master entity
 		ConditionModel<Attribute<?>, Entity> masterModel =
 						model.attribute(Detail.MASTER_FK);